ijepa_loader_optimized.py
#!/usr/bin/env python3
"""
Optimized I-JEPA Model Loader
=============================

This module provides an optimized implementation for loading and using I-JEPA models
with efficient feature extraction for YOLOv8 integration.

Key optimizations:
- Cached model loading
- Efficient feature extraction
- Memory optimization
- Batch processing support
"""

import torch
import torch.nn as nn
from transformers import AutoImageProcessor, AutoModel
import torchvision.transforms as transforms
from PIL import Image
import numpy as np
import time
from typing import Optional, Tuple, Union
import warnings

# Suppress warnings for cleaner output
warnings.filterwarnings("ignore")


class IJEPALoaderOptimized:
    """
    Optimized I-JEPA model loader with efficient feature extraction
    
    This class provides fast and memory-efficient loading and feature extraction
    from I-JEPA models for integration with YOLOv8.
    """
    
    def __init__(self, model_name: str = "facebook/ijepa_vith16_1k", 
                 device: Optional[str] = None,
                 cache_dir: Optional[str] = None):
        """
        Initialize optimized I-JEPA loader
        
        Args:
            model_name: Hugging Face model name
            device: Device to load model on (auto-detect if None)
            cache_dir: Directory to cache model files
        """
        self.model_name = model_name
        self.device = device or ('cuda' if torch.cuda.is_available() else 'cpu')
        self.cache_dir = cache_dir
        
        # Model and processor will be loaded lazily
        self.model = None
        self.processor = None
        self._is_loaded = False
        
        # Feature cache for repeated extractions
        self._feature_cache = {}
        self._cache_size = 100  # Maximum cache size
        
        print(f"🔧 Initialized I-JEPA loader on {self.device}")
    
    def load_and_freeze(self) -> Tuple[nn.Module, AutoImageProcessor]:
        """
        Load I-JEPA model and processor, then freeze parameters
        
        Returns:
            Tuple of (model, processor)
        """
        if self._is_loaded:
            return self.model, self.processor
        
        print(f"Loading I-JEPA model: {self.model_name}")
        start_time = time.time()
        
        try:
            # Load model with optimizations
            self.model = AutoModel.from_pretrained(
                self.model_name,
                cache_dir=self.cache_dir,
                torch_dtype=torch.float16 if self.device == 'cuda' else torch.float32,
                low_cpu_mem_usage=True
            )
            
            # Load processor
            self.processor = AutoImageProcessor.from_pretrained(
                self.model_name,
                cache_dir=self.cache_dir
            )
            
            # Move to device
            self.model = self.model.to(self.device)
            
            # Freeze all parameters for efficiency
            for param in self.model.parameters():
                param.requires_grad = False
            
            # Set to evaluation mode
            self.model.eval()
            
            # Enable optimizations
            if self.device == 'cuda':
                self.model = torch.compile(self.model, mode="reduce-overhead")
            
            self._is_loaded = True
            
            load_time = time.time() - start_time
            print(f"Model loaded successfully in {load_time:.2f}s")
            
        except Exception as e:
            print(f"Error loading model: {e}")
            raise
        
        return self.model, self.processor
    
    def extract_features_optimized(self, images: torch.Tensor) -> torch.Tensor:
        """
        Extract features from images using optimized I-JEPA
        
        Args:
            images: Input tensor [batch_size, 3, height, width]
            
        Returns:
            Feature tensor [batch_size, feature_dim]
        """
        if not self._is_loaded:
            self.load_and_freeze()
        
        # Move images to device
        images = images.to(self.device)
        
        # Create cache key
        cache_key = hash(images.cpu().numpy().tobytes())
        
        # Check cache first
        if cache_key in self._feature_cache:
            return self._feature_cache[cache_key]
        
        # Preprocess images efficiently
        processed_images = self._preprocess_images_optimized(images)
        
        # Extract features
        with torch.no_grad():
            try:
                outputs = self.model(**processed_images)
                
                # Extract features based on model output structure
                if hasattr(outputs, 'last_hidden_state'):
                    features = outputs.last_hidden_state
                elif hasattr(outputs, 'hidden_states') and outputs.hidden_states is not None:
                    features = outputs.hidden_states[-1]
                else:
                    features = outputs[0] if isinstance(outputs, tuple) else outputs
                
                # Global average pooling if needed
                if features.dim() == 3:
                    features = features.mean(dim=1)
                
                # Cache the result
                if len(self._feature_cache) < self._cache_size:
                    self._feature_cache[cache_key] = features
                
                return features
                
            except Exception as e:
                print(f"Error extracting features: {e}")
                # Fallback to simpler processing
                return self._extract_features_fallback(images)
    
    def _preprocess_images_optimized(self, images: torch.Tensor) -> dict:
        """
        Optimized image preprocessing for I-JEPA
        
        Args:
            images: Input tensor [batch_size, 3, height, width]
            
        Returns:
            Processed inputs dictionary
        """
        batch_size = images.shape[0]
        processed_images = []
        
        # Convert to PIL images efficiently
        for i in range(batch_size):
            # Convert tensor to PIL image
            img_tensor = images[i]
            img_array = (img_tensor.cpu().numpy().transpose(1, 2, 0) * 255).astype(np.uint8)
            pil_image = Image.fromarray(img_array)
            
            # Process with I-JEPA processor
            processed = self.processor(pil_image, return_tensors="pt")
            processed_images.append(processed['pixel_values'].squeeze(0))
        
        # Stack processed images
        batch_inputs = {
            'pixel_values': torch.stack(processed_images).to(self.device)
        }
        
        return batch_inputs
    
    def _extract_features_fallback(self, images: torch.Tensor) -> torch.Tensor:
        """
        Fallback feature extraction method
        
        Args:
            images: Input tensor [batch_size, 3, height, width]
            
        Returns:
            Feature tensor [batch_size, feature_dim]
        """
        # Simple feature extraction as fallback
        # This is a simplified version for when the main extraction fails
        batch_size = images.shape[0]
        
        # Use a simple CNN-like feature extraction
        features = torch.mean(images, dim=[2, 3])  # Global average pooling
        features = torch.relu(features)  # Apply ReLU
        
        # Project to expected feature dimension
        if features.shape[1] != 1280:  # I-JEPA feature dimension
            projection = nn.Linear(features.shape[1], 1280).to(self.device)
            features = projection(features)
        
        return features
    
    def extract_features_batch(self, image_paths: list, batch_size: int = 8) -> torch.Tensor:
        """
        Extract features from a list of image paths in batches
        
        Args:
            image_paths: List of image file paths
            batch_size: Batch size for processing
            
        Returns:
            Feature tensor [num_images, feature_dim]
        """
        if not self._is_loaded:
            self.load_and_freeze()
        
        all_features = []
        
        # Process in batches
        for i in range(0, len(image_paths), batch_size):
            batch_paths = image_paths[i:i + batch_size]
            batch_images = []
            
            # Load batch images
            for img_path in batch_paths:
                try:
                    image = Image.open(img_path).convert('RGB')
                    # Convert to tensor
                    transform = transforms.ToTensor()
                    img_tensor = transform(image)
                    batch_images.append(img_tensor)
                except Exception as e:
                    print(f"Warning: Could not load {img_path}: {e}")
                    # Add zero tensor as placeholder
                    batch_images.append(torch.zeros(3, 224, 224))
            
            # Stack batch
            if batch_images:
                batch_tensor = torch.stack(batch_images)
                features = self.extract_features_optimized(batch_tensor)
                all_features.append(features)
        
        # Concatenate all features
        if all_features:
            return torch.cat(all_features, dim=0)
        else:
            return torch.empty(0, 1280)
    
    def get_feature_dimension(self) -> int:
        """Get the feature dimension of the loaded model"""
        if not self._is_loaded:
            self.load_and_freeze()
        
        # Test with dummy input to get feature dimension
        dummy_input = torch.randn(1, 3, 224, 224).to(self.device)
        features = self.extract_features_optimized(dummy_input)
        return features.shape[1]
    
    def clear_cache(self):
        """Clear the feature cache"""
        self._feature_cache.clear()
        print("🗑️ Feature cache cleared")
    
    def get_model_info(self) -> dict:
        """Get information about the loaded model"""
        if not self._is_loaded:
            return {"status": "not_loaded"}
        
        info = {
            "model_name": self.model_name,
            "device": self.device,
            "feature_dim": self.get_feature_dimension(),
            "cache_size": len(self._feature_cache),
            "parameters": sum(p.numel() for p in self.model.parameters()),
            "trainable_parameters": sum(p.numel() for p in self.model.parameters() if p.requires_grad)
        }
        
        return info


def test_optimized_loader():
    """Test the optimized I-JEPA loader"""

    # Create loader
    loader = IJEPALoaderOptimized()
    
    # Load model
    model, processor = loader.load_and_freeze()
    
    # Test feature extraction
    print("Testing feature extraction...")
    
    # Test with different input sizes
    test_sizes = [(1, 3, 224, 224), (2, 3, 640, 640), (4, 3, 448, 448)]
    
    for batch_size, channels, height, width in test_sizes:
        print(f"\nTesting batch size: {batch_size}, size: {height}x{width}")
        
        # Create dummy input
        dummy_input = torch.randn(batch_size, channels, height, width)
        
        # Time the feature extraction
        start_time = time.time()
        features = loader.extract_features_optimized(dummy_input)
        extraction_time = time.time() - start_time
        
        print(f"Features shape: {features.shape}")
        print(f"⏱Extraction time: {extraction_time:.3f}s")
        print(f"Speed: {batch_size/extraction_time:.1f} images/second")
    
    # Get model info
    info = loader.get_model_info()
    print(f"\n📊 Model Info:")
    for key, value in info.items():
        print(f"   {key}: {value}")
    
    print("\n✅ Optimized loader test completed!")


if __name__ == "__main__":
    test_optimized_loader() 