requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
ultralytics>=8.0.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0
jupyter>=1.0.0
ipywidgets>=7.6.0
timm>=0.6.0
albumentations>=1.1.0
PyYAML>=6.0
requests>=2.25.0 